Renombramientos y movimiento de archivos para cumplir la especificación TMDBRemoteResult.kt -> RemoteResult.kt                     -> tmdb/RemoteMovie.kt OMDBRemoteResult.kt -> omdb/RemoteMovie.kt
