+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/external/tmdb/RemoteResult.kt
@@ -1,0 +1,15 @@
+package edu.dyds.movies.data.external.tmdb
+
+import edu.dyds.movies.domain.entity.Movie
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TMDBRemoteResult(
+    val page: Int,
+    val results: List<RemoteMovie>,
+    @SerialName("total_pages") val totalPages: Int,
+    @SerialName("total_results") val totalResults: Int
+)
+
+
