--- composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/external/tmdb/TMDBRemoteResult.kt
@@ -1,42 +1,0 @@
-package edu.dyds.movies.data.external.tmdb
-
-import edu.dyds.movies.domain.entity.Movie
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TMDBRemoteResult(
-    val page: Int,
-    val results: List<TMDBRemoteMovie>,
-    @SerialName("total_pages") val totalPages: Int,
-    @SerialName("total_results") val totalResults: Int
-)
-
-@Serializable
-data class TMDBRemoteMovie(
-    val id: Int,
-    val title: String,
-    val overview: String,
-    @SerialName("release_date") val releaseDate: String?,
-    @SerialName("poster_path") val posterPath: String?,
-    @SerialName("backdrop_path") val backdropPath: String?,
-    @SerialName("original_title") val originalTitle: String,
-    @SerialName("original_language") val originalLanguage: String,
-    val popularity: Double?,
-    @SerialName("vote_average") val voteAverage: Double?,
-)
-
-fun TMDBRemoteMovie.TMDBtoDomainMovie(): Movie {
-    return Movie(
-        id = id,
-        title = title,
-        overview = overview,
-        releaseDate = releaseDate?: "",
-        poster = "https://image.tmdb.org/t/p/w185$posterPath",
-        backdrop = backdropPath.let { "https://image.tmdb.org/t/p/w780$it" },
-        originalTitle = originalTitle,
-        originalLanguage = originalLanguage,
-        popularity = popularity ?: 0.0,
-        voteAverage = voteAverage ?: 0.0
-    )
-}
